convert_and_confirm_modal is migrated
